requirements.txt
requests
flask
gunicorn
flask_sqlalchemy
pymysql
python_gnupg
werkzeug